# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/problem-add": {
            "post": {
                "description": "添加问题",
                "tags": [
                    "管理员私有方法"
                ],
                "summary": "添加一个问题",
                "parameters": [
                    {
                        "type": "string",
                        "description": "问题标题",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "问题内容",
                        "name": "content",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "最大内存",
                        "name": "max_mem",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "最大运行时间",
                        "name": "max_runtime",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "description": "分类id",
                        "name": "category_ids",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "description": "测试用例",
                        "name": "test_cases",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{“code”: \"200\", \"msg\":\"\", \"data\": \"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/problem/detail": {
            "get": {
                "description": "获取问题详细信息",
                "tags": [
                    "公共方法"
                ],
                "summary": "问题详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "问题的唯一标识",
                        "name": "identity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{“code”: \"200\", \"msg\":\"\", \"data\": \"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/problem/list": {
            "get": {
                "description": "获取问题列表",
                "tags": [
                    "公共方法"
                ],
                "summary": "获取问题列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "请输入当前页,默认第一页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页多少条数据,默认20条",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "查询的关键字",
                        "name": "keyWord",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分类的唯一标识",
                        "name": "category_identity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{“code”: \"200\", \"msg\":\"\", \"data\": \"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}